+++ webapp/app/[locale]/tunnel/_components/withdraw.tsx
@@ -1,6 +1,6 @@
-import { useBridgeState } from 'app/[locale]/bridge/useBridgeState'
-import { useTransactionsList } from 'app/[locale]/bridge/useTransactionsList'
-import { useWithdraw } from 'app/[locale]/bridge/useWithdraw'
+import { useBridgeState } from 'app/[locale]/tunnel/useBridgeState'
+import { useTransactionsList } from 'app/[locale]/tunnel/useTransactionsList'
+import { useWithdraw } from 'app/[locale]/tunnel/useWithdraw'
 import { useNativeTokenBalance, useTokenBalance } from 'hooks/useBalance'
 import dynamic from 'next/dynamic'
 import { useTranslations } from 'next-intl'
