+++ webapp/components/header.tsx
@@ -8,6 +8,7 @@ import Skeleton from 'react-loading-skeleton'
 import { HamburgerIcon } from 'ui-common/components/hamburgerIcon'
 import { HemiLogoFull, HemiSymbol } from 'ui-common/components/hemiLogo'
 import { useOnClickOutside } from 'ui-common/hooks/useOnClickOutside'
+import { useTranslations } from 'next-intl'
 
 const WalletConnectButton = dynamic(
   () =>
@@ -32,7 +33,7 @@ const WalletConnectMobile = dynamic(
 )
 
 type Props = {
-  path: '/bridge' | '/swap'
+  path: '/swap' | '/tunnel'
   text: string
 }
 
@@ -53,6 +54,7 @@ const Route = function ({ path, text }: Props) {
 
 export const Header = function () {
   const [isMenuOpen, setIsMenuOpen] = useState(false)
+  const t = useTranslations('common')
 
   const toggleMenu = () => setIsMenuOpen(!isMenuOpen)
 
@@ -71,8 +73,8 @@ export const Header = function () {
       <div className="hidden">
         <div className="fixed bottom-0 z-10 flex w-full pb-6 pt-4 md:absolute md:left-1/2 md:block md:-translate-x-1/2">
           <ul className="mx-auto hidden justify-center gap-x-4 rounded-xl bg-neutral-100 p-[2px] md:mx-0 md:bg-inherit">
-            <Route path="/bridge" text="Bridge" />
-            <Route path="/swap" text="Swap" />
+            <Route path="/tunnel" text={t('tunnel')} />
+            <Route path="/swap" text={t('swap')} />
           </ul>
         </div>
       </div>
